refactor(import): isolate object selection utility
Keep ImportService focused on orchestration and move typed search resolution into a dedicated module.
